+++ packages/node-rnw-rpc/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - node-rnw-rpc
 
-This log was last generated on Wed, 10 Mar 2021 05:10:47 GMT and should not be manually modified.
+This log was last generated on Fri, 12 Mar 2021 05:07:46 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.0.5-2
+
+Fri, 12 Mar 2021 05:07:46 GMT
+
+### Changes
+
+- Bump react-native-windows to v0.0.0-canary.272 (ngerlem@microsoft.com)
+
+## 1.0.5-1
+
+Thu, 11 Mar 2021 05:07:32 GMT
+
+### Changes
+
+- Bump react-native-windows to v0.0.0-canary.271 (asklar@microsoft.com)
 
 ## 1.0.5-0